Configuración de la inyección de dependencia en la clases DAO y controladores, se editó archivo appsetting.json, se modifica clase startup para inyectar las dependencias, validaciones en formularios al agregar o editar

+++ WebBodega/WebBodega/Pages/Reporte/IndexReporte.cshtml.cs
@@ -30,7 +30,7 @@ namespace WebBodega
         public async Task OnPostReporte()
         {
             var model = Servicio;
-            int idCliente = model.IdCliente;
+            int idCliente = (int)model.IdCliente;
             var httpClient = new HttpClient();
             var json = await httpClient.GetStringAsync($"https://localhost:44351/api/servicioalojamiento/reporte/{idCliente}");
             Producto = JsonConvert.DeserializeObject<List<ProductoModel>>(json);
